# Use traditional pointer approach for RTree