--- a/rtree/bulk.go
+++ b/rtree/bulk.go
@@ -19,9 +19,6 @@ func BulkLoad(items []BulkItem) RTree {
 }
 
 func (t *RTree) bulkInsert(items []BulkItem) *node {
-	if len(items) == 0 {
-		return nil
-	}
 	if len(items) <= 2 {
 		root := &node{isLeaf: true, numEntries: len(items)}
 		for i, item := range items {