--- a/rtree/rtree.go
+++ b/rtree/rtree.go
@@ -7,18 +7,29 @@ import (
 // node is a node in an R-Tree. nodes can either be leaf nodes holding entries
 // for terminal items, or intermediate nodes holding entries for more nodes.
 type node struct {
-	isLeaf  bool
-	entries []entry
-	parent  int
+	entries    [1 + maxChildren]entry
+	numEntries int
+	parent     *node
+	isLeaf     bool
 }
 
 // entry is an entry under a node, leading either to terminal items, or more nodes.
 type entry struct {
 	box Box
 
-	// For leaf nodes, this is a record ID. For non-leaf nodes, this is the
-	// index of the child node.
-	index int
+	// For leaf nodes, recordID is populated. For non-leaf nodes, child is populated.
+	child    *node
+	recordID int
+}
+
+func (n *node) appendRecord(box Box, recordID int) {
+	n.entries[n.numEntries] = entry{box: box, recordID: recordID}
+	n.numEntries++
+}
+
+func (n *node) appendChild(box Box, child *node) {
+	n.entries[n.numEntries] = entry{box: box, child: child}
+	n.numEntries++
 }
 
 // RTree is an in-memory R-Tree data structure. It holds record ID and bounding
@@ -26,8 +37,7 @@ type entry struct {
 // responsible for storing their own records). Its zero value is an empty
 // R-Tree.
 type RTree struct {
-	rootIndex int
-	nodes     []node
+	root *node
 }
 
 // Stop is a special sentinal error that can be used to stop a search operation
@@ -41,37 +51,38 @@ var Stop = errors.New("stop")
 // case where the special Stop sentinal error is returned (in which case nil
 // will be returned from Search).
 func (t *RTree) Search(box Box, callback func(recordID int) error) error {
-	if len(t.nodes) == 0 {
+	if t.root == nil {
 		return nil
 	}
 	var recurse func(*node) error
 	recurse = func(n *node) error {
-		for _, entry := range n.entries {
+		for i := 0; i < n.numEntries; i++ {
+			entry := n.entries[i]
 			if !overlap(entry.box, box) {
 				continue
 			}
 			if n.isLeaf {
-				if err := callback(entry.index); err == Stop {
+				if err := callback(entry.recordID); err == Stop {
 					return nil
 				} else if err != nil {
 					return err
 				}
 			} else {
-				if err := recurse(&t.nodes[entry.index]); err != nil {
+				if err := recurse(entry.child); err != nil {
 					return err
 				}
 			}
 		}
 		return nil
 	}
-	return recurse(&t.nodes[t.rootIndex])
+	return recurse(t.root)
 }
 
 // Extent gives the Box that most closely bounds the RTree. If the RTree is
 // empty, then false is returned.
 func (t *RTree) Extent() (Box, bool) {
-	if len(t.nodes) == 0 {
+	if t.root == nil || t.root.numEntries == 0 {
 		return Box{}, false
 	}
-	return t.calculateBound(t.rootIndex), true
+	return calculateBound(t.root), true
 }