## Description

The existing implementation uses an array/index based approach. There
was originally an intention that the data structure would be able to be
serialised to disk, hence an array/index based approach was used. That
seems a bit too specialised, and that approach was becoming very hard to
work with (e.g. to implement delete).

Using a more tradition node/pointer based approach allows easier
implementation of new extensions to the RTree (delete and ranged search
are the main ones in mind, but there may be more in the future).

## Check List

Have you:

- Added unit tests? Yes

- Add cmprefimpl tests? (if appropriate?) N/A

## Related Issue

- Relates to #194 (although doesn't address it).

## Benchmark Results

There is a decent performance increase for operations that use the rtree (that wasn't really intentional, but not going to complain about it).

```
name                                          old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4       3.80µs ± 9%    2.32µs ±14%  -39.01%  (p=0.000 n=14+13)
IntersectsLineStringWithLineString/n=100-4      53.9µs ± 9%    40.2µs ±10%  -25.29%  (p=0.000 n=14+15)
IntersectsLineStringWithLineString/n=1000-4      732µs ± 4%     607µs ±13%  -17.11%  (p=0.000 n=13+15)
IntersectsLineStringWithLineString/n=10000-4    12.7ms ±13%     9.8ms ±17%  -22.40%  (p=0.000 n=15+14)
LineStringIsSimpleZigZag/10-4                   3.86µs ± 7%    2.16µs ± 6%  -44.05%  (p=0.000 n=14+14)
LineStringIsSimpleZigZag/100-4                  99.5µs ±11%    64.1µs ±11%  -35.57%  (p=0.000 n=15+15)
LineStringIsSimpleZigZag/1000-4                 1.36ms ± 6%    1.01ms ± 3%  -26.10%  (p=0.000 n=14+14)
LineStringIsSimpleZigZag/10000-4                18.1ms ± 4%    14.1ms ± 8%  -22.23%  (p=0.000 n=12+15)
PolygonSingleRingValidation/n=10-4              4.62µs ±16%    2.48µs ±17%  -46.38%  (p=0.000 n=14+14)
PolygonSingleRingValidation/n=100-4              100µs ± 8%      58µs ± 8%  -42.15%  (p=0.000 n=15+14)
PolygonSingleRingValidation/n=1000-4            1.48ms ± 7%    1.08ms ± 6%  -27.03%  (p=0.000 n=15+15)
PolygonSingleRingValidation/n=10000-4           22.3ms ±16%    14.5ms ± 4%  -34.98%  (p=0.000 n=14+14)
PolygonMultipleRingsValidation/n=4-4            16.0µs ± 7%    10.7µs ± 6%  -33.40%  (p=0.000 n=14+13)
PolygonMultipleRingsValidation/n=36-4            145µs ± 6%     100µs ± 8%  -31.19%  (p=0.000 n=15+14)
PolygonMultipleRingsValidation/n=400-4          1.88ms ± 7%    1.26ms ± 4%  -32.89%  (p=0.000 n=15+12)
PolygonMultipleRingsValidation/n=4096-4         20.8ms ± 3%    15.1ms ± 9%  -27.44%  (p=0.000 n=13+15)
PolygonZigZagRingsValidation/n=10-4             36.4µs ±10%    26.5µs ± 8%  -27.30%  (p=0.000 n=14+14)
PolygonZigZagRingsValidation/n=100-4             457µs ± 2%     345µs ± 6%  -24.48%  (p=0.000 n=13+13)
PolygonZigZagRingsValidation/n=1000-4           6.02ms ± 9%    4.87ms ±13%  -19.11%  (p=0.000 n=14+15)
PolygonZigZagRingsValidation/n=10000-4          79.3ms ± 3%    63.8ms ± 3%  -19.57%  (p=0.000 n=13+13)
MultipolygonValidation/n=1-4                     387ns ± 7%     356ns ±14%   -8.05%  (p=0.002 n=13+14)
MultipolygonValidation/n=4-4                    1.07µs ± 6%    0.92µs ±14%  -14.11%  (p=0.000 n=14+14)
MultipolygonValidation/n=16-4                   10.0µs ±14%     5.8µs ± 8%  -42.30%  (p=0.000 n=14+14)
MultipolygonValidation/n=64-4                   61.0µs ± 8%    34.8µs ± 8%  -42.92%  (p=0.000 n=14+15)
MultipolygonValidation/n=256-4                   327µs ±15%     208µs ± 9%  -36.34%  (p=0.000 n=14+14)
MultipolygonValidation/n=1024-4                 1.66ms ±12%    1.03ms ± 8%  -37.88%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=10-4                   10.5µs ± 7%     7.7µs ±15%  -26.43%  (p=0.000 n=13+14)
MultiPolygonTwoCircles/n=100-4                   140µs ± 8%     108µs ± 8%  -22.95%  (p=0.000 n=15+14)
MultiPolygonTwoCircles/n=1000-4                 1.97ms ±28%    1.57ms ±14%  -20.34%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=10000-4                27.5ms ±20%    24.0ms ±26%  -12.72%  (p=0.001 n=14+14)
MultiPolygonMultipleTouchingPoints/n=1-4        7.84µs ±10%    6.05µs ± 9%  -22.88%  (p=0.000 n=13+14)
MultiPolygonMultipleTouchingPoints/n=10-4       63.6µs ±12%    52.2µs ±12%  -17.87%  (p=0.000 n=14+15)
MultiPolygonMultipleTouchingPoints/n=100-4       733µs ± 7%     613µs ± 4%  -16.34%  (p=0.000 n=14+14)
MultiPolygonMultipleTouchingPoints/n=1000-4     9.19ms ± 5%    7.86ms ± 4%  -14.46%  (p=0.000 n=13+12)
MarshalWKB/polygon/n=10-4                        203ns ±11%     209ns ±21%     ~     (p=0.517 n=13+15)
MarshalWKB/polygon/n=100-4                       597ns ±22%     611ns ±17%     ~     (p=0.382 n=14+13)
MarshalWKB/polygon/n=1000-4                     4.00µs ±17%    4.27µs ±35%     ~     (p=0.511 n=13+13)
MarshalWKB/polygon/n=10000-4                    33.5µs ±35%    33.2µs ±54%     ~     (p=0.683 n=15+13)
UnmarshalWKB/polygon/n=10-4                      343ns ±10%     342ns ±19%     ~     (p=0.705 n=15+15)
UnmarshalWKB/polygon/n=100-4                     729ns ±14%     725ns ±17%     ~     (p=0.765 n=14+13)
UnmarshalWKB/polygon/n=1000-4                   4.14µs ±22%    4.38µs ±21%     ~     (p=0.098 n=13+12)
UnmarshalWKB/polygon/n=10000-4                  41.0µs ±40%    40.3µs ±31%     ~     (p=0.769 n=14+14)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                             52.0µs ±12%    51.9µs ±10%     ~     (p=0.847 n=14+15)
Intersection/n=100-4                            98.3µs ± 7%    95.3µs ± 4%   -3.05%  (p=0.044 n=14+14)
Intersection/n=1000-4                            431µs ±13%     442µs ± 5%     ~     (p=0.128 n=14+13)
Intersection/n=10000-4                          3.68ms ±14%    3.77ms ±12%     ~     (p=0.345 n=15+15)
NoOp/n=10-4                                     3.94µs ± 6%    3.94µs ± 9%     ~     (p=0.949 n=15+14)
NoOp/n=100-4                                    12.8µs ± 3%    12.9µs ± 7%     ~     (p=0.693 n=11+15)
NoOp/n=1000-4                                   91.3µs ± 6%    91.7µs ± 8%     ~     (p=0.839 n=14+14)
NoOp/n=10000-4                                   975µs ±23%     995µs ±15%     ~     (p=0.461 n=15+15)

name                                          old alloc/op   new alloc/op   delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4       3.63kB ± 0%    2.69kB ± 0%  -25.99%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=100-4      40.1kB ± 0%    32.6kB ± 0%  -18.60%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=1000-4      369kB ± 0%     286kB ± 0%  -22.35%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=10000-4    5.45MB ± 0%    3.77MB ± 0%  -30.87%  (p=0.000 n=14+13)
LineStringIsSimpleZigZag/10-4                   3.34kB ± 0%    1.44kB ± 0%  -56.94%  (p=0.000 n=15+15)
LineStringIsSimpleZigZag/100-4                  64.4kB ± 0%    21.6kB ± 0%  -66.43%  (p=0.000 n=14+15)
LineStringIsSimpleZigZag/1000-4                  667kB ± 0%     230kB ± 0%  -65.58%  (p=0.000 n=14+12)
LineStringIsSimpleZigZag/10000-4                7.33MB ± 0%    2.30MB ± 0%  -68.58%  (p=0.000 n=14+13)
PolygonSingleRingValidation/n=10-4              3.47kB ± 0%    1.47kB ± 0%  -57.60%  (p=0.000 n=15+15)
PolygonSingleRingValidation/n=100-4             62.4kB ± 0%    16.2kB ± 0%  -74.12%  (p=0.000 n=15+15)
PolygonSingleRingValidation/n=1000-4             666kB ± 0%     217kB ± 0%  -67.39%  (p=0.000 n=15+15)
PolygonSingleRingValidation/n=10000-4           7.34MB ± 0%    2.23MB ± 0%  -69.68%  (p=0.000 n=12+15)
PolygonMultipleRingsValidation/n=4-4            7.90kB ± 0%    5.28kB ± 0%  -33.20%  (p=0.000 n=15+15)
PolygonMultipleRingsValidation/n=36-4           73.1kB ± 0%    43.3kB ± 0%  -40.72%  (p=0.000 n=15+14)
PolygonMultipleRingsValidation/n=400-4           847kB ± 0%     476kB ± 0%  -43.81%  (p=0.000 n=14+12)
PolygonMultipleRingsValidation/n=4096-4         9.08MB ± 0%    4.88MB ± 0%  -46.20%  (p=0.000 n=12+14)
PolygonZigZagRingsValidation/n=10-4             21.2kB ± 0%    14.0kB ± 0%  -34.14%  (p=0.000 n=15+15)
PolygonZigZagRingsValidation/n=100-4             247kB ± 0%     138kB ± 0%  -44.09%  (p=0.000 n=13+15)
PolygonZigZagRingsValidation/n=1000-4           2.38MB ± 0%    1.26MB ± 0%  -46.95%  (p=0.000 n=15+14)
PolygonZigZagRingsValidation/n=10000-4          30.5MB ± 0%    15.3MB ± 0%  -49.62%  (p=0.000 n=13+14)
MultipolygonValidation/n=1-4                      225B ± 0%      385B ± 0%  +71.11%  (p=0.000 n=15+15)
MultipolygonValidation/n=4-4                      820B ± 0%      676B ± 0%  -17.56%  (p=0.000 n=15+15)
MultipolygonValidation/n=16-4                   8.53kB ± 0%    3.86kB ± 0%  -54.78%  (p=0.000 n=15+15)
MultipolygonValidation/n=64-4                   44.7kB ± 0%    15.4kB ± 0%  -65.49%  (p=0.000 n=14+15)
MultipolygonValidation/n=256-4                   185kB ± 0%      63kB ± 0%  -65.80%  (p=0.000 n=14+15)
MultipolygonValidation/n=1024-4                  783kB ± 0%     259kB ± 0%  -66.96%  (p=0.000 n=15+12)
MultiPolygonTwoCircles/n=10-4                   7.81kB ± 0%    6.88kB ± 0%  -11.88%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=100-4                  74.4kB ± 0%    59.4kB ± 0%  -20.16%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=1000-4                  673kB ± 0%     508kB ± 0%  -24.52%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=10000-4                10.3MB ± 0%     6.9MB ± 0%  -32.83%  (p=0.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1-4        4.79kB ± 0%    4.18kB ± 0%  -12.70%  (p=0.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=10-4       26.9kB ± 0%    23.8kB ± 0%  -11.55%  (p=0.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=100-4       243kB ± 0%     212kB ± 0%  -12.63%  (p=0.000 n=14+15)
MultiPolygonMultipleTouchingPoints/n=1000-4     2.44MB ± 0%    2.00MB ± 0%  -18.01%  (p=0.000 n=15+14)
MarshalWKB/polygon/n=10-4                         232B ± 0%      232B ± 0%     ~     (all equal)
MarshalWKB/polygon/n=100-4                      1.83kB ± 0%    1.83kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=1000-4                     16.4kB ± 0%    16.4kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=10000-4                     164kB ± 0%     164kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10-4                       282B ± 0%      282B ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=100-4                    1.90kB ± 0%    1.90kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=1000-4                   16.5kB ± 0%    16.5kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10000-4                   164kB ± 0%     164kB ± 0%     ~     (all equal)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                             1.19kB ± 0%    1.19kB ± 0%     ~     (all equal)
Intersection/n=100-4                            6.33kB ± 0%    6.33kB ± 0%     ~     (all equal)
Intersection/n=1000-4                           55.0kB ± 0%    55.0kB ± 0%     ~     (p=0.156 n=12+15)
Intersection/n=10000-4                           558kB ± 0%     558kB ± 0%     ~     (p=0.871 n=14+15)
NoOp/n=10-4                                       864B ± 0%      864B ± 0%     ~     (all equal)
NoOp/n=100-4                                    5.68kB ± 0%    5.68kB ± 0%     ~     (all equal)
NoOp/n=1000-4                                   49.5kB ± 0%    49.5kB ± 0%     ~     (all equal)
NoOp/n=10000-4                                   492kB ± 0%     492kB ± 0%     ~     (p=0.068 n=15+14)

name                                          old allocs/op  new allocs/op  delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
IntersectsLineStringWithLineString/n=10-4         34.0 ± 0%      15.0 ± 0%  -55.88%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=100-4         363 ± 0%       160 ± 0%  -55.92%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=1000-4      3.06k ± 0%     1.28k ± 0%  -58.14%  (p=0.000 n=15+15)
IntersectsLineStringWithLineString/n=10000-4     33.6k ± 0%     19.3k ± 0%  -42.69%  (p=0.000 n=15+15)
LineStringIsSimpleZigZag/10-4                     27.0 ± 0%       5.0 ± 0%  -81.48%  (p=0.000 n=15+15)
LineStringIsSimpleZigZag/100-4                     441 ± 0%        75 ± 0%  -82.99%  (p=0.000 n=15+15)
LineStringIsSimpleZigZag/1000-4                  4.62k ± 0%     0.80k ± 0%  -82.74%  (p=0.000 n=15+15)
LineStringIsSimpleZigZag/10000-4                 46.4k ± 0%      8.0k ± 0%  -82.76%  (p=0.000 n=15+15)
PolygonSingleRingValidation/n=10-4                30.0 ± 0%       6.0 ± 0%  -80.00%  (p=0.000 n=15+15)
PolygonSingleRingValidation/n=100-4                427 ± 0%        57 ± 0%  -86.65%  (p=0.000 n=15+15)
PolygonSingleRingValidation/n=1000-4             4.61k ± 0%     0.76k ± 0%  -83.63%  (p=0.000 n=15+15)
PolygonSingleRingValidation/n=10000-4            46.5k ± 0%      7.7k ± 0%  -83.37%  (p=0.000 n=15+15)
PolygonMultipleRingsValidation/n=4-4              90.0 ± 0%      25.0 ± 0%  -72.22%  (p=0.000 n=15+15)
PolygonMultipleRingsValidation/n=36-4              778 ± 0%       203 ± 0%  -73.91%  (p=0.000 n=15+15)
PolygonMultipleRingsValidation/n=400-4           8.89k ± 0%     2.23k ± 0%  -74.94%  (p=0.000 n=15+15)
PolygonMultipleRingsValidation/n=4096-4          92.1k ± 0%     22.8k ± 0%  -75.19%  (p=0.000 n=15+14)
PolygonZigZagRingsValidation/n=10-4                209 ± 0%        70 ± 0%  -66.51%  (p=0.000 n=15+15)
PolygonZigZagRingsValidation/n=100-4             2.00k ± 0%     0.64k ± 0%  -68.23%  (p=0.000 n=15+15)
PolygonZigZagRingsValidation/n=1000-4            18.4k ± 0%      5.5k ± 0%  -70.37%  (p=0.000 n=15+15)
PolygonZigZagRingsValidation/n=10000-4            194k ± 0%       74k ± 0%  -61.88%  (p=0.000 n=15+15)
MultipolygonValidation/n=1-4                      6.00 ± 0%      5.00 ± 0%  -16.67%  (p=0.000 n=15+15)
MultipolygonValidation/n=4-4                      11.0 ± 0%       8.0 ± 0%  -27.27%  (p=0.000 n=15+15)
MultipolygonValidation/n=16-4                     68.0 ± 0%      27.0 ± 0%  -60.29%  (p=0.000 n=15+15)
MultipolygonValidation/n=64-4                      325 ± 0%        99 ± 0%  -69.54%  (p=0.000 n=15+15)
MultipolygonValidation/n=256-4                   1.33k ± 0%     0.39k ± 0%  -70.44%  (p=0.000 n=15+15)
MultipolygonValidation/n=1024-4                  5.57k ± 0%     1.58k ± 0%  -71.64%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=10-4                     86.0 ± 0%      46.0 ± 0%  -46.51%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=100-4                     736 ± 0%       326 ± 0%  -55.71%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=1000-4                  6.13k ± 0%     2.57k ± 0%  -58.11%  (p=0.000 n=15+15)
MultiPolygonTwoCircles/n=10000-4                 67.3k ± 0%     38.6k ± 0%  -42.68%  (p=0.000 n=14+15)
MultiPolygonMultipleTouchingPoints/n=1-4          75.0 ± 0%      51.0 ± 0%  -32.00%  (p=0.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=10-4          432 ± 0%       332 ± 0%  -23.15%  (p=0.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=100-4       3.88k ± 0%     3.07k ± 0%  -20.87%  (p=0.000 n=12+14)
MultiPolygonMultipleTouchingPoints/n=1000-4      36.3k ± 0%     29.2k ± 0%  -19.57%  (p=0.000 n=15+15)
MarshalWKB/polygon/n=10-4                         6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=100-4                        6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=1000-4                       6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=10000-4                      6.00 ± 0%      6.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10-4                       7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=100-4                      7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=1000-4                     7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10000-4                    7.00 ± 0%      7.00 ± 0%     ~     (all equal)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
Intersection/n=10-4                               31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=100-4                              31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=1000-4                             31.0 ± 0%      31.0 ± 0%     ~     (all equal)
Intersection/n=10000-4                            31.0 ± 0%      31.0 ± 0%     ~     (all equal)
NoOp/n=10-4                                       21.0 ± 0%      21.0 ± 0%     ~     (all equal)
NoOp/n=100-4                                      21.0 ± 0%      21.0 ± 0%     ~     (all equal)
NoOp/n=1000-4                                     21.0 ± 0%      21.0 ± 0%     ~     (all equal)
NoOp/n=10000-4                                    21.0 ± 0%      21.0 ± 0%     ~     (all equal)
```